+++ CHANGELOG.md
@@ -45,6 +45,7 @@ Contains all the PRs that improved the code without changing the behaviors.
 - Fixed validator rewards payout.
 - Fixed bug allowing double claims with Thorchain claims.
 - Fixed non deterministic map iteration to sorted iteration 
+- Fixed fee stuck in arkeo module by moving fees to arkeo-reserve
 
 ---